+++ daspi/__init__.py
@@ -309,7 +309,7 @@ from .plotlib.plotter import Pareto
 from .plotlib.plotter import Jitter
 from .plotlib.plotter import Plotter
 from .plotlib.plotter import Scatter
-from .plotlib.plotter import Violine
+from .plotlib.plotter import Violin
 from .plotlib.plotter import Beeswarm
 from .plotlib.plotter import Errorbar
 from .plotlib.plotter import MeanTest